requirements.txt
paho-mqtt
numpy
scipy
sounddevice
